# swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the ORY Hydra HTTP API documentation. You will find documentation for all HTTP APIs here.",
    "title": "ORY Hydra",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/identifier/{id}": {
      "get": {
        "description": "Delete the data identifier with the identifierID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "dataIdentifier"
        ],
        "summary": "Delete a data identifier",
        "operationId": "dataIdentifierID",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/jsonError"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      },
      "delete": {
        "description": "Delete the data identifier with the identifierID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "dataIdentifier"
        ],
        "summary": "Delete a data identifier",
        "operationId": "dataIdentifierID",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/jsonError"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      }
    },
...
    "/subscriptions": {
      "post": {
        "description": "Create a new data identifier subscription if you have the valid license(apiKey). The privateKey and publicKey are returned in response.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "subscription"
        ],
        "summary": "Create a data identifier subscription",
        "operationId": "createSubscription",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string",
                  "x-go-name": "Apikey"
                },
                "identifier": {
                  "type": "string",
                  "x-go-name": "Identifier"
                },
                "recipient": {
                  "type": "string",
                  "x-go-name": "Recipient"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/subscriptionResp"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      }
    },
    "/subscriptions/{id}": {
      "get": {
        "description": "Get the data identifier subscription with the subscriptionID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "subscription"
        ],
        "summary": "Get the data identifier subscription",
        "operationId": "dataSubscriptionID",
        "responses": {
          "200": {
            "$ref": "#/responses/subscriptionResp"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      },
      "delete": {
        "description": "Delete the data identifier subscription with the subscriptionID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "subscription"
        ],
        "summary": "Delete the data identifier subscription",
        "operationId": "dataSubscriptionID",
        "responses": {
          "200": {
            "$ref": "#/responses/jsonError"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      },
      "patch": {
        "description": "Audit the data identifier subscription with the subscriptionID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "subscription"
        ],
        "summary": "Audit the data identifier subscription",
        "operationId": "auditSubscription",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "x-go-name": "Status"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/subscriptionResp"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Metadata": {
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/ory/hydra/subscription"
    },
...
    "Subscription": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "expired_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ExpiredAt"
        },
        "id": {
          "description": "Hash(Requestor+Target+Owner).Target",
          "type": "string",
          "x-go-name": "ID"
        },
        "identifier": {
          "type": "string",
          "x-go-name": "Identifier"
        },
        "metadata": {
          "$ref": "#/definitions/Metadata"
        },
        "modified_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ModifiedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "recipient": {
          "type": "string",
          "x-go-name": "Recipient"
        },
        "requestor": {
          "type": "string",
          "x-go-name": "Requestor"
        },
        "status": {
          "$ref": "#/definitions/SubscriptionStatus"
        },
        "type": {
          "$ref": "#/definitions/SubscriptionType"
        }
      },
      "x-go-package": "github.com/ory/hydra/subscription"
    },
    "SubscriptionStatus": {
      "type": "string",
      "x-go-package": "github.com/ory/hydra/subscription"
    },
    "SubscriptionType": {
      "type": "string",
      "x-go-package": "github.com/ory/hydra/subscription"
    },
...
  },
  "responses": {
...
    "subscriptionResp": {
      "description": "Subscription information response are sent when the operation succeeds.",
      "schema": {
        "$ref": "#/definitions/Subscription"
      }
    }
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    },
    "oauth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://hydra.demo.ory.sh/oauth2/auth",
      "tokenUrl": "https://hydra.demo.ory.sh/oauth2/token",
      "scopes": {
        "offline": "A scope required when requesting refresh tokens (alias for `offline_access`)",
        "offline_access": "A scope required when requesting refresh tokens",
        "openid": "Request an OpenID Connect ID Token"
      }
    }
  },
...
}